Move `pmap` executor to its own mod
